core: Cleanup of attio_offline list
This patch removes the attio_offline list used to notify the ATT
connections when the ATTIO callback is registered. If a client registers
dynamically the callbacks, the ATT channel could be established
previouslly triggered by other clients. For this scenario, the GAttrib
instance needs to be informed to the client without triggering scanning
or connection.